pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testforge"
version = "1.0.0"
description = "Comprehensive mutation testing framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TestForge Team", email = "dev@testforge.io"}
]
keywords = [
    "mutation-testing",
    "test-quality",
    "coverage",
    "testing",
    "software-quality",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.8"

dependencies = [
    "astor>=0.8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "flake8>=6.0",
]
all = [
    "testforge[dev]",
    "coverage>=7.0",
]

[project.scripts]
testforge = "testforge.cli:main"

[project.urls]
Homepage = "https://github.com/moggan1337/TestForge"
Documentation = "https://testforge.readthedocs.io"
Repository = "https://github.com/moggan1337/TestForge"
Issues = "https://github.com/moggan1337/TestForge/issues"
Changelog = "https://github.com/moggan1337/TestForge/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["testforge", "testforge.core", "testforge.operators", "testforge.analysis", "testforge.integration", "testforge.reporting", "testforge.debugging", "testforge.autogen", "testforge.cicd", "testforge.utils"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
testforge = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
